add http:// prefix to a few links
Some of the links returned 404 error messages due to the rst used. Rst assumes a link without a prefix is contained in the local directory, though none of the links in this file are. 

For example, the line 

* `The Atlas of Economic Complexity <atlas.cid.harvard.edu>`_

would proceed to the url https://github.com/caesar0301/awesome-public-datasets/blob/master/atlas.cid.harvard.edu, resulting in a 404 error. 

My change prepends http:// to the link so that line now routes to the correct address.
New line: 

* `The Atlas of Economic Complexity <http://atlas.cid.harvard.edu>`_