Add new open dataset - 2016 New Coder Survey
This dataset contains more than 15,000 records of responses to 48 questions from people who are new to software development. It was conducted by FreeCodeCamp.com and CodeNewbie.org with the intention as serving as an objective, open dataset for use by researchers, journalists, and employers. We have both the original datasets and a cleaned and unified dataset, all under the Open Database License.